Basic Path From Screen
Just started to implement the basic path creation from Screen. The idea
is to just simply consume step by step until we have no more valid
steps. The output should be a list of steps we have to do to simply walk
on the path from the beginning to the end.

Example:
[Left, Forward, Forward, Right, Forward]