Jump back on the project
I still want to learn and improve and hopefully one day I can be a decent
Haskell developer.

First step now, to configure vscode and try if it works or not. If not,
I will go back to Vim as I wrote the rest. So brittany configures with
ghcide. hie is installed also, but now let's stick with ghcide. I don't
know what's the real difference and which one is better.

That's the main reason why set back the resolver to lts-14.27 because
ghcide did not like ghc-8.8x :(